Struct positioned_io::Cursor

pub struct Cursor<I> {
    // some fields omitted
}

Adapts a ReadAt or WriteAt into a Read or Write.

This wraps anything that read and write at offsets, turning into an object that can read or write at a file position. This allows you to use those types with all the many functions that expect a Read or Write.

Note that seeking on Cursor has limited functionality. We don't know how many bytes are available, so we can't use SeekFrom::End. See SizeCursor for another option.

Examples

use positioned_io::{ReadAt, Cursor};

struct NetworkStorage {
    // A remote disk that supports random access.
}
impl ReadAt for NetworkStorage {
    // ...
}

// Adapt our storage into a Read.
let storage = NetworkStorage::new(SOME_LOCATION);
let curs = Cursor::new_pos(storage, 1 << 30);

// Copy a segment to a file.
let mut input = curs.take(1 << 20);
let mut output = try!(File::create("segment.out"));
try!(io::copy(&mut input, &mut output));

Methods

impl<I> Cursor<I>

fn new_pos(io: I, pos: u64) -> Self

Create a new Cursor which starts reading at a specified offset.

Pass in a ReadAt or WriteAt as io.

fn new(io: I) -> Self

Create a new Cursor which starts reading at offset zero.

Pass in a ReadAt or WriteAt as io.

fn into_inner(self) -> I

Consume self and yield the inner ReadAt or WriteAt.

fn get_ref(&self) -> &I

Borrow the inner ReadAt or WriteAt.

fn get_mut(&mut self) -> &mut I

Borrow the inner ReadAt or WriteAt mutably.

fn position(&self) -> u64

Get the current read/write position.

fn set_position(&mut self, pos: u64)

Set the current read/write position.