Struct positioned_io::RandomAccessFile

pub struct RandomAccessFile { /* private fields */ }

A wrapper for File that provides optimized random access through ReadAt and WriteAt.

On Unix the operating system is advised that reads will be in random order (FADV_RANDOM).

On Windows the implementation is orders of magnitude faster than ReadAt directly on File.

Examples

Read the fifth 512-byte sector of a file:

use positioned_io::{RandomAccessFile, ReadAt};

// open a file (note: binding does not need to be mut)
let raf = RandomAccessFile::open("tests/pi.txt")?;

// read up to 512 bytes
let mut buf = [0; 512];
let bytes_read = raf.read_at(2048, &mut buf)?;

Implementations

impl RandomAccessFile

pub fn open<P: AsRef<Path>>(path: P) -> Result<RandomAccessFile>

Opens a file for random access.

pub fn try_new(file: File) -> Result<RandomAccessFile>

Creates a RandomAccessFile wrapper around a File.

pub fn try_into_inner(self) -> Result<File, (RandomAccessFile, Error)>

Tries to unwrap the inner File.