Struct positioned_io::SizeCursor

pub struct SizeCursor<I: Size>(_);

Adapts a ReadAt or WriteAt into a Read or Write, with better seeking.

This is just like Cursor, except that it requires an object that implements Size, and that it can seek from the end of the I/O object.

Eventually it will be legal to specialize Cursor for types that implement Size, see RFC 1210.

Methods

impl<I> SizeCursor<I> where I: Size

fn new_pos(io: I, pos: u64) -> Self

Create a new SizeCursor which starts reading at a specified offset.

Pass in a ReadAt or WriteAt as io.

fn new(io: I) -> Self

Create a new SizeCursor which starts reading at offset zero.

Pass in a ReadAt or WriteAt as io.