Struct positioned_io::Slice

pub struct Slice<I> { /* private fields */ }

A window into another ReadAt or WriteAt.

Given an existing positioned I/O, this presents a limited view of it.

Examples

Some slices have size restrictions:

use positioned_io::{ReadAt, Slice};

let a = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
let slice = Slice::new(&a[..], 4, Some(4));

let mut buf = [0; 4];
let bytes = slice.read_at(2, &mut buf)?;
assert_eq!(bytes, 2);
assert_eq!(buf, [6, 7, 0, 0]);

Some slices do not:

use positioned_io::{WriteAt, Slice};

let mut v = vec![0, 1, 2, 3, 4, 5];
let buf = [9; 3];

{
    let mut slice = Slice::new(&mut v, 2, None);
    slice.write_all_at(3, &buf)?;
}

// The write goes right past the end.
assert_eq!(v, vec![0, 1, 2, 3, 4, 9, 9, 9]);

Implementations

impl<I> Slice<I>

pub fn new(io: I, offset: u64, size: Option<u64>) -> Self

Create a new Slice.

The slice will be a view of size bytes, starting at offset in io. If you do not pass a size, the size won’t be limited.

pub fn into_inner(self) -> I

Consumes the slice, returning the underlying value.

pub fn get_ref(&self) -> &I

Get a reference to the underlying value in this slice.

Note that IO on the returned value may escape the slice.

pub fn get_mut(&mut self) -> &mut I

Get a mutable reference to the underlying value in this slice.

Note that IO on the returned value may escape the slice.

pub fn offset(&self) -> u64

Get the offset that this slice starts at within the underlying IO.

pub fn set_offset(&mut self, offset: u64)

Set the offset that this slice starts at within the underlying IO.

impl<I: Size> Slice<I>

pub fn new_to_end(io: I, offset: u64) -> Result<Self>

Create a new Slice that goes to the end of io.

Note that you can create a larger slice by passing a larger size to new(), but it won’t do you any good for reading.