Struct positioned_io::ByteIo

pub struct ByteIo<I, E: ByteOrder> { /* private fields */ }

Read or write with a given inherent byte-order.

If you know that you’ll always be using a single endianness, an instance of ByteIo will allow you to omit the endian specifier on every read or write.

Examples

use byteorder::BigEndian;
use positioned_io::ByteIo;

let mut buf = [0; 8];

{
    let mut io : ByteIo<_, BigEndian> = ByteIo::new(&mut buf[..]);
    // All writes will automatically be BigEndian.
    io.write_u16(300)?;
    io.write_u32(1_000_000)?;
    io.write_i16(-1)?;
}

assert_eq!(buf, [1, 44, 0, 15, 66, 64, 255, 255]);

Implementations

impl<I, E: ByteOrder> ByteIo<I, E>

pub fn new(io: I) -> Self

Create a new ByteIo from some sort of reader or writer.

You will need to specify the byte-order when creating a ByteIo.

Examples

use byteorder::BigEndian;
use positioned_io::ByteIo;

let buf = [0; 10];
// Add a type specifier for the byte order.
let io : ByteIo<_, BigEndian> = ByteIo::new(&buf[..]);

pub fn into_inner(self) -> I

Returns the underlying reader or writer.

pub fn get_ref(&self) -> &I

Borrows the underlying reader or writer.

pub fn get_mut(&mut self) -> &mut I

Mutably borrows the underlying reader or writer.

impl<I: Read, E: ByteOrder> ByteIo<I, E>

pub fn read_u8(&mut self) -> Result<u8>

pub fn read_i8(&mut self) -> Result<i8>

pub fn read_u16(&mut self) -> Result<u16>

pub fn read_i16(&mut self) -> Result<i16>

pub fn read_u32(&mut self) -> Result<u32>

pub fn read_i32(&mut self) -> Result<i32>

pub fn read_u64(&mut self) -> Result<u64>

pub fn read_i64(&mut self) -> Result<i64>

pub fn read_uint(&mut self, nbytes: usize) -> Result<u64>

pub fn read_int(&mut self, nbytes: usize) -> Result<i64>

pub fn read_f32(&mut self) -> Result<f32>

pub fn read_f64(&mut self) -> Result<f64>

impl<I: Write, E: ByteOrder> ByteIo<I, E>

pub fn write_u8(&mut self, n: u8) -> Result<()>

pub fn write_i8(&mut self, n: i8) -> Result<()>

pub fn write_u16(&mut self, n: u16) -> Result<()>

pub fn write_i16(&mut self, n: i16) -> Result<()>

pub fn write_u32(&mut self, n: u32) -> Result<()>

pub fn write_i32(&mut self, n: i32) -> Result<()>

pub fn write_u64(&mut self, n: u64) -> Result<()>

pub fn write_i64(&mut self, n: i64) -> Result<()>

pub fn write_uint(&mut self, n: u64, nbytes: usize) -> Result<()>

pub fn write_int(&mut self, n: i64, nbytes: usize) -> Result<()>

pub fn write_f32(&mut self, n: f32) -> Result<()>

pub fn write_f64(&mut self, n: f64) -> Result<()>

impl<I: ReadAt, E: ByteOrder> ByteIo<I, E>

pub fn read_u8_at(&self, pos: u64) -> Result<u8>

pub fn read_i8_at(&self, pos: u64) -> Result<i8>

pub fn read_u16_at(&self, pos: u64) -> Result<u16>

pub fn read_i16_at(&self, pos: u64) -> Result<i16>

pub fn read_u32_at(&self, pos: u64) -> Result<u32>

pub fn read_i32_at(&self, pos: u64) -> Result<i32>

pub fn read_u64_at(&self, pos: u64) -> Result<u64>

pub fn read_i64_at(&self, pos: u64) -> Result<i64>

pub fn read_uint_at(&self, pos: u64, nbytes: usize) -> Result<u64>

pub fn read_int_at(&self, pos: u64, nbytes: usize) -> Result<i64>

pub fn read_f32_at(&self, pos: u64) -> Result<f32>

pub fn read_f64_at(&self, pos: u64) -> Result<f64>

impl<I: WriteAt, E: ByteOrder> ByteIo<I, E>

pub fn write_u8_at(&mut self, pos: u64, n: u8) -> Result<()>

pub fn write_i8_at(&mut self, pos: u64, n: i8) -> Result<()>

pub fn write_u16_at(&mut self, pos: u64, n: u16) -> Result<()>

pub fn write_i16_at(&mut self, pos: u64, n: i16) -> Result<()>

pub fn write_u32_at(&mut self, pos: u64, n: u32) -> Result<()>

pub fn write_i32_at(&mut self, pos: u64, n: i32) -> Result<()>

pub fn write_u64_at(&mut self, pos: u64, n: u64) -> Result<()>

pub fn write_i64_at(&mut self, pos: u64, n: i64) -> Result<()>

pub fn write_uint_at(&mut self, pos: u64, n: u64, nbytes: usize) -> Result<()>

pub fn write_int_at(&mut self, pos: u64, n: i64, nbytes: usize) -> Result<()>

pub fn write_f32_at(&mut self, pos: u64, n: f32) -> Result<()>

pub fn write_f64_at(&mut self, pos: u64, n: f64) -> Result<()>